backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for Hackathon Todo"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlmodel",
    "psycopg2-binary",
    "python-jose[cryptography]",
    "python-dotenv",
    "asyncpg",
    "google-generativeai>=0.8.6",
]
